+++ Core/Changelog/7.2.4.lua
@@ -3,19 +3,15 @@ local TXUI, F, E, I, V, P, G = unpack((select(2, ...)))
 TXUI.Changelog["7.2.4"] = {
   HOTFIX = true,
   CHANGES = {
-    "* Breaking changes",
-
     "* New features",
-    "Replace the " .. TXUI.Title .. " installer's Turbo Mode with Import Existing, where it finds existing profiles and suggests using the newest one" .. F.String.Sublist(
-      "This should make setting up alts easier"
-    ) .. F.String.Sublist("Please note that this feature does not include any other addon profiles, only the ElvUI profiles"),
+    "Replace the "
+      .. TXUI.Title
+      .. " installer's Turbo Mode with Import Existing, where it finds existing profiles and suggests using the newest one" --
+      .. F.String.Sublist("This should make setting up alts easier")
+      .. F.String.Sublist("Please note that this feature does not include any other addon profiles, only the ElvUI profiles"),
 
     "* Enhancements",
     F.String.Retail() .. "Display mana percentage for healers in " .. F.String.ToxiUI("[tx:power]") .. " tag",
-
-    "* Bug fixes",
-
-    "* Profile updates",
 
     "* Documentation",
     "Add " .. F.String.Epic("evilknivel") .. " to the contributors list",