+++ CHANGELOG.md
@@ -1,3 +1,24 @@
 # ToxiUI v7.2.4 - 2026-02-05
 
-_crickets_
+## 🚀 New Features
+
+- Replace the **ToxiUI** installer’s **Turbo Mode** with **Import Existing**, which finds existing profiles and suggests using the newest one
+  - This should make setting up alts easier
+  - This feature does **not** include any other addon profiles, only ElvUI profiles
+
+## ✨ Enhancements
+
+- [Retail] Display mana percentage for healers in the **[tx:power]** tag
+
+## 📘 Documentation
+
+- Add **evilknivel** to the contributors list
+- Add **Exodus** to the contributors list
+
+## ♻️ Settings Refactoring
+
+- Add more spacing between changelog categories for improved readability
+
+## 🧰 Development Improvements
+
+- Create a table for Retail Specialization IDs for easier referencing